# intermediate/fitness-workout-tracker/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Exercise:
    properties:
      category:
        description: e.g., Strength, Cardio, Flexibility
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      muscleGroup:
        description: e.g., Chest, Legs, Back
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8800
info:
  contact:
    email: jaygaha@gmail.com
    name: API Support
    url: https://jaygaha.com.np
  description: This is a sample server fitness workout tracker server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fitness Workout Tracker API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user with email and password
      operationId: loginUser
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Email and password are required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password
      operationId: registerUser
      parameters:
      - description: Register Request
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully. Please log in to continue.
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Username or email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /exercises:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all available exercises
      operationId: getExercises
      produces:
      - application/json
      responses:
        "200":
          description: List of exercises
          schema:
            items:
              $ref: '#/definitions/models.Exercise'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      - BearerAuth: []
      summary: Get all exercises
      tags:
      - exercises
...
securityDefinitions:
  BearerAuth:
    description: 'Provide JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"